requirements.txt
asgiref==3.8.1
cffi==1.17.1
chardet==5.2.0
cryptography==43.0.3
Django==5.1.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.7
pillow==11.0.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.0
python-dotenv==1.0.1
reportlab==4.2.5
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
